% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Domenico.R
\name{const_Dom}
\alias{const_Dom}
\alias{square_Dom}
\title{Domenico solution}
\usage{
const_Dom(C0 = 1, x, y, z, t, vx, ax, ay, az, lambda, sY, sZ, Rf = 1,
  decaysorbed = TRUE)

square_Dom(C0 = 1, x, y, z, t, vx, ax, ay, az, lambda, sY, sZ, dur)
}
\arguments{
\item{C0}{source concentration}

\item{x, y, z}{evaluation locations}

\item{t}{evaluation time}

\item{vx}{velocity in x direction (it is assumed that velocity only has
an x component); this can be divided by the retardation factor to model
sorption}

\item{ax, ay, az}{dispersivities (units of length) in x, y and z
directions}

\item{lambda}{first-order decay constant}

\item{sY, sZ}{source dimensions in y and z directions}

\item{dur}{duration of pulse}
}
\value{
Numeric result, of same form as x, y or z.  x, y and z may be recycled.
}
\description{
The Domenico 1987 analytical solution to contaminant transport in
 uniform groundwater flow, including longitudinal and transverse
 dispersion and first-order degradation.
}
\details{
\code{const_Dom} returns the solution from a constant source which
 starts at time 0.  \code{square_Dom} returns the solution from a finite
 duration constant source that starts at time 0 and stops after
 \code{dur}.
}
\note{
Some special cases may be achieved with certain parameter sets.

If lambda = 0 (no degradation), the Domenico-Robbins (1985) solution
 results.\cr
If ax = 0 (no longitudinal dispersion), the Domenico-Palciauskas (1982)
 solution results.\cr
If ay = az = 0 (no transverse dispersion), the Bear (1979) solution
 results.\cr
If ay = az = 0 and lambda = 0, the Ogata-Banks (1961) solution
 results.\cr
}
\references{
Domenico, P. A. (1987). An analytical model for multidimensional transport of a decaying contaminant species. Journal of Hydrology, 91(1), 49-58. https://doi.org/10.1016/0022-1694(87)90127-2
}

